+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - **Other Features**
   - [spiral/filters] Added `Spiral\Filter\ValidationHandlerMiddleware` for handling filter validation exception.
   - [spiral/router] Fixed the problem with parsing a pattern with `0` value in route parameter.
+  - [spiral/validation] Added the ability to configure the default validator via method `setDefaultValidator` 
+    in the `Spiral\Validation\Bootloader\ValidationBootloader`.
 
 ## v3.0.0 - 2022-09-13
 - **High Impact Changes**